Changing entity id data types from int to Long

+++ app/src/main/java/com/matilfa/twentyquestions/data/questions/Question.java
@@ -10,7 +10,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Question {
     @PrimaryKey(autoGenerate = true)
-    public int questionId;
+    public Long questionId;
 
     @ColumnInfo(name = "text")
     public String text;
